+++ integration-tests/src/test/java/org/apache/druid/tests/security/ITTLSTest.java
@@ -29,6 +29,8 @@ import org.apache.druid.java.util.common.ISE;
 import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.java.util.common.lifecycle.Lifecycle;
 import org.apache.druid.java.util.common.logger.Logger;
+import org.apache.druid.java.util.emitter.core.NoopEmitter;
+import org.apache.druid.java.util.emitter.service.ServiceEmitter;
 import org.apache.druid.java.util.http.client.CredentialedHttpClient;
 import org.apache.druid.java.util.http.client.HttpClient;
 import org.apache.druid.java.util.http.client.HttpClientConfig;
@@ -392,7 +394,8 @@ public class ITTLSTest
 
     HttpClient client = HttpClientInit.createClient(
         builder.build(),
-        lifecycle
+        lifecycle,
+        new ServiceEmitter("", "", new NoopEmitter())
     );
 
     HttpClient adminClient = new CredentialedHttpClient(
@@ -419,7 +422,8 @@ public class ITTLSTest
 
     HttpClient client = HttpClientInit.createClient(
         builder.build(),
-        lifecycle
+        lifecycle,
+        new ServiceEmitter("", "", new NoopEmitter())
     );
 
     HttpClient adminClient = new CredentialedHttpClient(
