+++ server/src/test/java/org/apache/druid/server/initialization/JettyCertRenewTest.java
@@ -35,6 +35,7 @@ import org.apache.druid.guice.LifecycleModule;
 import org.apache.druid.guice.annotations.Self;
 import org.apache.druid.initialization.Initialization;
 import org.apache.druid.java.util.emitter.core.NoopEmitter;
+import org.apache.druid.java.util.emitter.service.ServiceEmitter;
 import org.apache.druid.java.util.http.client.HttpClient;
 import org.apache.druid.java.util.http.client.HttpClientConfig;
 import org.apache.druid.java.util.http.client.HttpClientInit;
@@ -387,7 +388,7 @@ public class JettyCertRenewTest extends BaseJettyTest
       client = HttpClientInit.createClient(
           getSslConfig(),
           lifecycle,
-          new NoopEmitter()
+          new ServiceEmitter("", "", new NoopEmitter())
       );
     }
     catch (Exception e) {
